添加多选菜单,完善展示流程.

+++ CollectionViewClassifyMenu/Class/Other/Category/NSArray+log.m
@@ -11,7 +11,7 @@
 @implementation NSArray (log)
 - (NSString *)descriptionWithLocale:(id)locale
 {
-    NSMutableString *str = [NSMutableString stringWithFormat:@"%d (\n", (int)self.count];
+    NSMutableString *str = [NSMutableString stringWithFormat:@"%d (\n", (NSUInteger)self.count];
     
     for (id obj in self) {
         [str appendFormat:@"\t%@,\n", obj];
